Fixed bugs #3
