--- src/features/StreetcodePage/ArtGalleryBlock/ArtGalleryItem/ArtGalleryItem.styles.scss
@@ -1,41 +1,0 @@
-@use "@sass/_utils.functions.scss" as *;
-@use "@sass/variables/_variables.colors.scss" as c;
-@use "@sass/mixins/_utils.mixins.scss" as mut;
-@use "@sass/mixins/_wrapper.mixins.scss" as wr;
-@use "@sass/mixins/_vendor.mixins.scss" as vnd;
-
-.slideArt {
-  flex: pxToRem(80px);
-  height: pxToRem(560px);
-}
-
-.imgWrap {
-  position: relative;  
-}
-
-.imgData {
-  @include mut.sized(100%, 113px);
-  @include mut.positioned-as($position: absolute, $bottom: 0);
-  @include mut.colored($color: c.$pure-white-color, $bg-color: rgba(0, 0, 0, 0.5));
-  visibility: hidden;
-  opacity: 0;
-  transition:opacity 1s; 
-  @include mut.rounded( $bottom-right: 30px, $bottom-left: 30px);
-  padding: pxToRem(11px) pxToRem(21px);
-} 
-
-.imgWrap:hover {
-  .imgDataFull {
-    visibility: visible;
-    opacity: 1;
-  }
-}
-
-.imgTitle {
-  @include mut.with-font();
-  @include mut.rem-padded($bottom: 5px);
-}
-
-.imgDescription {
-  @include mut.with-font($font-size: 15px);
-}
