+++ src/features/StreetcodePage/ArtGalleryBlock/ArtGalleryItem/ArtGallerySlideItem.component.tsx
@@ -1,4 +1,4 @@
-import './ArtGalleryItem.styles.scss';
+import './ArtGallerySlideItem.styles.scss';
 
 import { IndexedArt } from '@models/media/art.model';
 
@@ -6,15 +6,15 @@ interface Props {
   artGalleryItem: IndexedArt;
 }
 
-const ArtGalleryItem = ({ artGalleryItem }: Props) => {
+const ArtGallerySlideItem = ({ artGalleryItem }: Props) => {
     const { imageHref, description, title } = artGalleryItem;
     return (
         <div className="slideArt">
-            <div className="imgWrap">
+            <div className="artImageWrapper">
                 <img className="imgImg" src={imageHref} alt="" />
                 <div
                     className={`imgData imgData${
-                        !description && !title ? 'Empty' : 'Full'
+                        description || title ? 'Full' : 'Empty'
                     }`}
                 >
                     <p className="imgTitle">{title}</p>
@@ -25,4 +25,4 @@ const ArtGalleryItem = ({ artGalleryItem }: Props) => {
     );
 };
 
-export default ArtGalleryItem;
+export default ArtGallerySlideItem;
